src/pyproject.toml
[tool.poetry]
name = "consolida-pagamentos-job"
version = "0.1.0"
description = "Refactory of ConsolidaPagamentos.php"
authors = ["rickslayer <paulo.rube@madeiramadeira.com.br>"]

[tool.poetry.dependencies]
python = "^3.8"
pymysql = "^0.9.3"
boto3 = "^1.12.46"
pandas = "^1.0.3"
starlette = "^0.13.3"
uvicorn = "^0.11.5"
python-multipart = "^0.0.5"
newrelic = "^5.12.0"
requests = "^2.23.0"
xlrd = "^1.2.0"
httpie = "^2.1.0"
python-dotenv = "^0.14.0"
botocore = "^1.18.1"
numpy = "^1.19.2"
redis = "^3.5.3"
DateTime = "^4.3"
pytz = "^2020.5"
chardet = "^4.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest = "^5.4.1"
pytest-dependency = "^0.5.1"
coverage = "^5.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
